wip: add blog suggestions
